add landing page and interactive board views
add landing page with links to manuals and interactive board views. The
landing page is automatically generated from the filesystem structure.
Place a tumbnail.[jpg, jpeg, png, gif] inside the images folder and
optionaly an ibom.html at the root folder of the project to generate
output for it. The interactive board views can be generated with the
kicad plugin "InteractiveHTMlBom" [https://github.com/openscopeproject/InteractiveHtmlBom]

Signed-off-by: Thomas Schmid <tom@binary-kitchen.de>